SQUASH: tentatively cast const-ness away when calling free()
I _think_ the API to git_config_string() and git_config_pathmame()
that take "const char **dest" parameter and stores a pointer to an
allocated memory the caller is going to own in it is mistaken and
they should instead take "char **dest" to make it explicit that the
returned memory is now owned by the caller.  But for now, this would
be sufficient to squelch the compiler warnings on the new free().